qi-add-memory-testing-routines.patch
This adds the memory testing code to Qi.  It tests the range of memory with
several constants and then a 32-bit walking 1 pattern, and repeats forever.

The entire main SDRAM can be tested due to the fact this runs out of
steppingstone only and does not need to store anything outside of it.

It introduces a steppingstone-based stack for use entirely in steppingstone.

Signed-off-by: Andy Green <andy@openmoko.com>