qi-memory-test-when-no-kernels.patch
This makes a new behaviour for Qi, when there are no valid kernels then it
performs the memory test.  It means you can either move kernels out of the way
to get the memory test, or make an SD Card just with normal Qi on it and no
kernels to get a memory test.

Signed-off-by: Andy Green <andy@openmoko.com>