qi-gta02-compute-and-append-dynparts.patch
This adds dynamic partition computation support for Qi on GTA02.
It's for compatability with existing GTA02 using U-Boot partition
scheme where a bad block in the previous partition moves on the
start of the next partition by one block.

It's important that Qi has no private state, so we need to compute
the NAND arrangement each boot.  It turns out this is extremely fast
using Qi's bad block code originally from Xinagfu.

Signed-off-by: Andy Green <andy@openmoko.com>