qi-add-gta02-indentity-part-parsing.patch
Now everything else is in place, we are able to mount
the GTA02 "identity" partition and extract the USB Ethernet
MAC Address from it, and add it to the kernel commandline.

This causes the Ethernet gadget to use the same MAC address
each boot, simplifying DHCP server situation.  The MAC
address in the identity partition is globally unique from
the factory.

Signed-off-by: Andy Green <andy@openmoko.com>