include entire source tree in source package
It is probably worth spending the time to split them up with
multiple upstream tarballs, but for now, usbboot, xbboot, and so on
are in a single repository managed as a single project.

Teach debian/rules get-orig-source to build a single upstream tarball
with the entire source tree.

Include history of whole tree in the generated changelog.upstream.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>